refactor(core): split protocol into focused modules
